PUSH-0.2.5: ex03 is done leet correct.

+++ CPP-02/ex03/Point.hpp
@@ -6,7 +6,7 @@
 /*   By: hibouzid <hibouzid@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/13 15:13:45 by hibouzid          #+#    #+#             */
-/*   Updated: 2024/10/14 23:48:21 by hibouzid         ###   ########.fr       */
+/*   Updated: 2024/10/15 16:48:40 by hibouzid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,8 @@ class Point{
 		Point(float v_x, float v_y);
 		Point(const Point& copy_point);
 		Point& operator=(const Point &copy_point);
-		int get_x() const;
-		int get_y() const;
+		float get_x() const;
+		float get_y() const;
 		~Point();
 };
 bool bsp( Point const a, Point const b, Point const c, Point const point);
